backend_flask/requirements.txt
alembic==1.8.1
aniso8601==9.0.1
asgiref==3.6.0
attrs==22.1.0
autopep8==2.0.0
bidict==0.22.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Django==4.1.7
dnspython==1.16.0
eventlet==0.33.2
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-Migrate==4.0.0
flask-restx==1.0.3
Flask-SocketIO==5.3.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.1
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
packaging==22.0
pkgutil_resolve_name==1.3.10
pycodestyle==2.10.0
pypi-install==0.0.5
pyrsistent==0.19.2
python-decouple==3.6
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.6
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.44
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.10.0
